projects/week-2/twin/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.121.1",
    "openai>=2.7.2",
    "openai-agents>=0.5.1",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
